v0.0.14 - refactor sqlvalue
