+++ src/data_sources/sql/create_query/create_nested_query_recursive/mod.rs
@@ -5,9 +5,10 @@ use crate::{
     configuration::subgraph::{
         data_sources::sql::DialectEnum, entities::ServiceEntityConfig, SubGraphConfig,
     },
-    data_sources::sql::{SqlDataSource, SqlValue},
+    data_sources::sql::SqlDataSource,
     filter_operator::FilterOperator,
     resolver_type::ResolverType,
+    sql_value::SqlValue,
 };
 
 use super::JoinClauses;
