+++ src/data_sources/sql/services/update_one/mod.rs
@@ -3,7 +3,8 @@ use sqlx::Row;
 
 use crate::{
     configuration::subgraph::entities::ServiceEntityConfig,
-    data_sources::sql::{PoolEnum, SqlQuery, SqlValue},
+    data_sources::sql::{PoolEnum, SqlQuery},
+    sql_value::SqlValue,
 };
 
 use super::{ResponseRow, Services};
