+++ src/lib.rs
@@ -16,6 +16,7 @@ pub mod filter_operator;
 pub mod graphql;
 pub mod resolver_type;
 pub mod scalar_option;
+pub mod sql_value;
 pub mod utils;
 
 /// Starts the Subgraph Service. Initializes the DataSources and builds the GraphQL Schema.
